+++ docs/sphinxext/extra_types.py
@@ -1,5 +1,0 @@
-from __future__ import generator_stop
-
-
-def setup(app):
-    app.add_description_unit("2to3fixer", "2to3fixer", "%s (2to3 fixer)")